data_analytics/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
blinker==1.9.0
bokeh==3.4.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.0
Django==4.2.16
django-allauth==0.52.0
django-cors-headers==4.2.0
django-plotly-dash==2.4.3
djangorestframework==3.15.2
dnspython==2.7.0
dpd-components==0.1.0
exceptiongroup==1.2.2
Flask==3.0.3
fonttools==4.54.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
kiwisolver==1.4.7
lazy_loader==0.4
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.9.2
msgpack==1.1.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.76.0
opencv-contrib-python==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pymongo==4.12.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9.1
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==9.0.0
tifffile==2024.8.30
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
Werkzeug==3.0.6
whitenoise==6.8.2
xyzservices==2024.9.0
zipp==3.20.2
